+++ examples/transmit_fixed_length.py
@@ -10,7 +10,7 @@ with cc1101.CC1101() as transceiver:
     transceiver.set_symbol_rate_baud(600)
     transceiver.set_sync_mode(cc1101.SyncMode.NO_PREAMBLE_AND_SYNC_WORD)
     transceiver.set_packet_length_mode(cc1101.PacketLengthMode.FIXED)
-    transceiver.set_packet_length(4)
+    transceiver.set_packet_length_bytes(4)
     transceiver.disable_checksum()
     print(transceiver)
     transceiver.transmit(b"\xff\x00\xaa\xff")